refactor: move fullscreen functions together in Item folder
