--- src/components/Item/isElementFullscreen.js
@@ -1,10 +1,0 @@
-import { getGridItemDomElementClassName } from '../../modules/getGridItemDomElementClassName.js'
-
-export const isElementFullscreen = (itemId) => {
-    const fullscreenElement =
-        document.fullscreenElement || document.webkitFullscreenElement
-
-    return fullscreenElement?.classList.contains(
-        getGridItemDomElementClassName(itemId)
-    )
-}
